about to finish remain with topup and withdraw testing

+++ models/VASService.model.go
@@ -16,7 +16,7 @@ type VASService struct {
 	NickName    string    `gorm:"<-;unique;not null;type:varchar(255)" json:"nickname"`
 	Description string    `gorm:"<-;not null;type:varchar(255)" json:"description"`
 	ProviderId  uuid.UUID `gorm:"<-;not null;type:uuid" json:"provider_id"`
-	Rebate      float32   `gorm:"<-;not null;type:float;" json:"rebate"`
+	Rebate      float64   `gorm:"<-;not null;type:float;" json:"rebate"`
 	Status      string    `gorm:"<-;not null;type:varchar(255);default:Active" json:"status"`
 	UpdatedAt   int64     `gorm:"autoUpdateTime" json:"updated_at"`
 	CreatedAt   int64     `gorm:"autoCreateTime" json:"created_at"`
